# Improve completer rendering performance