--- a/packages/completer/src/model.ts
+++ b/packages/completer/src/model.ts
@@ -8,6 +8,15 @@ import { CompletionHandler } from './handler';
 import { Completer } from './widget';
 
 /**
+ * Escape HTML by native means of the browser.
+ */
+function escapeHTML(text: string) {
+  const node = document.createElement('span');
+  node.textContent = text;
+  return node.innerHTML;
+}
+
+/**
  * An implementation of a completer model.
  */
 export class CompleterModel implements Completer.IModel {
@@ -166,7 +175,18 @@ export class CompleterModel implements Completer.IModel {
     if (query) {
       return this._markup(query);
     }
-    return this._completionItems;
+    return this._completionItems.map(item => {
+      const newItem = Object.assign({}, item);
+      const newLabel = escapeHTML(newItem.label);
+      if (newLabel != newItem.label) {
+        // preserve the old label for insertText if not defined
+        if (typeof newItem.insertText === 'undefined') {
+          newItem.insertText = newItem.label;
+        }
+        newItem.label = newLabel;
+      }
+      return newItem;
+    });
   }
 
   /**
@@ -345,13 +365,14 @@ export class CompleterModel implements Completer.IModel {
       // e.g. Given label `foo(b, a, r)` and query `bar`,
       // don't count parameters, `b`, `a`, and `r` as matches.
       const index = item.label.indexOf('(');
-      const prefix = index > -1 ? item.label.substring(0, index) : item.label;
+      let prefix = index > -1 ? item.label.substring(0, index) : item.label;
+      prefix = escapeHTML(prefix);
       let match = StringExt.matchSumOfSquares(prefix, query);
       // Filter non-matching items.
       if (match) {
         // Highlight label text if there's a match
         let marked = StringExt.highlight(
-          item.label,
+          escapeHTML(item.label),
           match.indices,
           Private.mark
         );