… docs/source/extension/extension_migration.rst
… galata/test/jupyterlab/collapsible-headings.test.ts-snapshots/add-header-below-01-jupyterlab-linux.png
… galata/test/jupyterlab/completer.test.ts
… galata/test/jupyterlab/notebooks/completer_panel.py
… packages/completer-extension/schema/tracker.json
… packages/completer/test/model.spec.ts
… packages/completer/test/widget.spec.ts
… packages/terminal-extension/src/index.ts
… packages/ui-components/src/hoverbox.ts
… packages/ui-components/style/hoverbox.css