## References

- fixes #13047 on the tokens benchmark:
   - x8.45 speedup in Chrome,
   - x4.79 speedup in Firefox
- fixes #5911
- fixes #9790
   - supersedes #10591
- fixes #12995

## Code changes

Performance:
- when many items are present (more than fits within hard-coded maximum completer height), pre-compute completer box size so that we can add `contain: strict` style drastically reducing time browsers take to layout completer
   - height is set to maximum height (if there are fewer items than max height, we are not in performance critical path)
   - width is estimated using the "widest" item; the widest item is found using a label length heuristic
      -  this heuristic can be replaced in third-party completers by providing optional `IRenderer.itemWidthHeuristic` 
- lazily populate the completer:
   - initially only as many items are rendered as needed to fill the first "page" (up to max height)
   - subsequent items are added in following frames in chunks
      - we dynamically choose the size of chunks so that they fit in the optimal frame rendering time of ~16ms
      - we avoid introducing any layout shift thanks to margin-bottom trick which sets appropriate height for listing scrollbars
- use `z-index` instead of `jp-mod-outofview` class for temporarily hiding hover box item, accelerating hover box positioning resolution by avoiding extra style re-computation
- use much faster browser-native sanitization method over custom sanitizer since we no longer need to allow `<mark>` tags which are now added later

UI-profiler benchmark comparison - 1000 tokens each 50 characters in text editor:

| browser | state | IQM [Q1 - Q3] | mean | min |
|---|---|---|---|--|
| Chrome | before | 461.0 [448.7 – 474.4] ms | 464.2 ms | 431.4 ms |
| Firefox | before | 269.2 [256.8 – 281.5] ms | 272 ms | 238.1 ms |
| Chrome | after | 54.5 [50.8 – 59.3] ms | 56.7 ms | 42.8 ms |
| Firefox | after | 56.2 [53.3 – 60.3] ms | 58.5 ms | 46.2 ms |

A real-word example (`import numpy as np; np.<tab>` in notebook):

| browser | state | IQM [Q1 - Q3] | mean | min |
|---|---|---|---|--|
| Chrome | before | 414.8 [398.6 – 452.9] ms | 426.2 ms | 374.5 ms |
| Firefox | before | 303.5 [283.3 – 338.9] ms | 320.9 ms | 261.2 ms |
| Chrome | after | 252.6 [235.2 – 277.7] ms | 255.6 ms | 212.5 ms |
| Firefox | after | 288.7 [252.3 – 320.5] ms | 302.6 ms | 230.8 ms |

Other bug-fixes/enhancements:
- properly escape completion items containing HTML or HTML-like code
- the default completer `Renderer` constructor signature was changed to follow `IOptions` pattern
- fixes layout of hover box when two panels are open one above the other (see `spaceAbove`/`spaceBelow` changes)
- remove `sanitizer` from field which was added to `IRenderer` interface in #13341: it is an implementation detail of our default `Renderer`, I think there is no need to require third-party renderers to expose their sanitizer
- re-align the completer widget `MAX_HEIGHT` (was 300 px) and `MIN_HEIGHT` (was 20px) with actual values defined in CSS styles (max-height was 240px, min-height was 24px) keeping CSS as dynamic source of truth (with cache); this means that changes made in themes will be properly reflected in positioning of completer hover box.

## User-facing changes

To allow dynamic (below/above cursor) placement of completer with pre-computed height we need to change the out-of-view behaviour of completer hover box. This means that the completer will no longer be hidden when its border first touch the parent (`hidden-inside` mode); instead it will be dynamically resized (using `stick-inside` mode); the alternative  would be to allow it to overflow slightly over the host node (using `hidden-outside` mode).


The default renderer/styles now use CSS `display: grid` over `display: table`; this is needed to prevent flickering with delayed rendering of invisible parts of completer. This results in smaller width of the completer in some circumstances (when pre-computed dimensions are used):

<details>

| before | after |
|---|---|
| ![Screenshot from 2022-12-27 14-47-14](https://user-images.githubusercontent.com/5832902/209725680-5d8c4599-e13a-4e5d-8c61-5728318afbe0.png) |![Screenshot from 2022-12-27 14-46-46](https://user-images.githubusercontent.com/5832902/209725692-ebce65cb-68ad-4725-9269-9f42340cfe67.png) |

Se how `VisibleDeprecationWarning` has just enough space to fit next to type information (`class`).

</details>

## Backwards-incompatible changes

- [x] `IRenderer.sanitizer` will be breaking unless we quickly remove it before 3.6 release
   - https://github.com/jupyterlab/jupyterlab/pull/13700 thanks @fcollonval
- removal of `jp-mod-outofview` class?