@krassowski 2022-12-27
please update galata snapshots